refactor: add winning lotto machine to LottoMachine file
